BUGFIX: too long titles in ipconfig
If the title is too long, it would fail the `'. :' in line` check, and that'd put the new title as an item in the previous title.
FIX: added another condition `or (not line.startswith("     ") and ':' in line)`
additionally, if the value is "", make it an empty list instead of a [""]
Upon adding an item to a list, if the list's new length is 1, change it back to a single item