BUGFIX in hostify
I had the arguments as `['nslookup','address']`.
This means the address is NOT passed in as an argument.
And I was fine with that?! It resolved to the socket.gethostbyaddr so there was data there but... c'mon, how did I not think to check which side is working?!

Anyway, now it's `['nslookup', address]` so the address is actually passed in as its value...