NetworkStorage._resolve hostify input addresses
Use hostify_sync to hostify and cache all addresses currently in the queue, then, if the entities entered have no name (entity.name == nothing.name), set their name to the hostify result.

This implementation allows for custom names upon input (in NetworkStorage.add), but when such names are not specified, it automatically tries to find them.

Note, however, that this is not perfectly optimised. If a scan knows it has a set of IP addresses, it SHOULD send that set to hostify/hostify_sync before `add`ing any entities, especially if the entities are not `add`ed simultaneously.