Increase repeats to 20, and remove infinite
scan_ICMP, scan_ARP, and scan_TCP, were all ramped up to 20 repeats, so that measurements can be more accurate. Inspiration was directly taken from Friday's physics experiment, in which we measured the time of 20 oscillations to minimise the error (when dividing by 20).
Removed continuous_ICMP from actions queue, because that would not go well with measuring time.