I first wrote this method
That allows getting the information about a specific scan. This choice of where to start was directly triggered by seeing the Information Button functionality (or lack thereof) in `gui.py`.

I searched how to protect against SQL Injections, if somehow possible in this system. The answer: use question marks (?), and pass the arguments in a list/tuple.